Extract @waveform-playlist/spectrogram package with extensible TrackMenu
Move spectrogram computation (FFT, window functions, color maps, frequency
scales) and SpectrogramSettingsModal into standalone package. Refactor
TrackMenu to accept generic items prop with onClose callback, making it
extensible for future menu items. Default MIR example to waveform mode
for faster page load with lazy spectrogram computation on mode switch.
