+++ packages/browser/src/index.tsx
@@ -135,9 +135,9 @@ export type { RenderAnnotationItemProps, AnnotationData } from '@waveform-playli
 // Export annotation callback types for Waveform components
 export type { GetAnnotationBoxLabelFn, OnAnnotationUpdateFn } from './types/annotations';
 
-// Export spectrogram computation and utilities
-export { computeSpectrogram, computeSpectrogramMono, getColorMap, getFrequencyScale } from './spectrogram';
-export type { FrequencyScaleName } from './spectrogram';
+// Re-export spectrogram computation and utilities
+export { computeSpectrogram, computeSpectrogramMono, getColorMap, getFrequencyScale } from '@waveform-playlist/spectrogram';
+export type { FrequencyScaleName } from '@waveform-playlist/spectrogram';
 
 // Export waveform-data.js utilities
 export {
