+++ packages/spectrogram/package.json
@@ -1,0 +1,56 @@
+{
+  "name": "@waveform-playlist/spectrogram",
+  "version": "5.2.0",
+  "description": "Spectrogram computation and UI for waveform-playlist",
+  "main": "./dist/index.js",
+  "module": "./dist/index.mjs",
+  "types": "./dist/index.d.ts",
+  "exports": {
+    ".": {
+      "types": "./dist/index.d.ts",
+      "import": "./dist/index.mjs",
+      "require": "./dist/index.js"
+    }
+  },
+  "sideEffects": false,
+  "scripts": {
+    "build": "tsup",
+    "dev": "tsup --watch",
+    "typecheck": "tsc --noEmit"
+  },
+  "keywords": [
+    "waveform",
+    "audio",
+    "spectrogram",
+    "fft",
+    "waveform-playlist"
+  ],
+  "author": "Naomi Aro",
+  "license": "MIT",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/naomiaro/waveform-playlist.git",
+    "directory": "packages/spectrogram"
+  },
+  "homepage": "https://naomiaro.github.io/waveform-playlist",
+  "bugs": {
+    "url": "https://github.com/naomiaro/waveform-playlist/issues"
+  },
+  "files": [
+    "dist",
+    "README.md"
+  ],
+  "devDependencies": {
+    "@types/react": "^18.2.45",
+    "@types/styled-components": "^5.1.26",
+    "tsup": "^8.0.1",
+    "typescript": "^5.3.3"
+  },
+  "dependencies": {
+    "@waveform-playlist/core": "workspace:*"
+  },
+  "peerDependencies": {
+    "react": "^18.0.0",
+    "styled-components": "^6.0.0"
+  }
+}
