+++ packages/spectrogram/src/components/index.ts
@@ -1,0 +1,5 @@
+export { SpectrogramMenuItems } from './SpectrogramMenuItems';
+export type { SpectrogramMenuItemsProps } from './SpectrogramMenuItems';
+export { SpectrogramSettingsModal } from './SpectrogramSettingsModal';
+export type { SpectrogramSettingsModalProps } from './SpectrogramSettingsModal';
+export type { TrackMenuItem } from './types';
