+++ packages/spectrogram/src/components/types.ts
@@ -1,0 +1,7 @@
+import type { ReactNode } from 'react';
+
+export interface TrackMenuItem {
+  id: string;
+  label?: string;
+  content: ReactNode;
+}
