+++ packages/spectrogram/src/computation/colorMaps.ts
