+++ packages/spectrogram/src/computation/computeSpectrogram.ts
