+++ packages/spectrogram/src/computation/frequencyScales.ts
