+++ packages/spectrogram/src/computation/index.ts
