+++ packages/spectrogram/src/computation/windowFunctions.ts
