+++ packages/spectrogram/src/index.ts
@@ -1,0 +1,10 @@
+// Computation
+export { computeSpectrogram, computeSpectrogramMono, getColorMap, getFrequencyScale } from './computation';
+export type { FrequencyScaleName } from './computation';
+
+// Components
+export { SpectrogramMenuItems } from './components';
+export type { SpectrogramMenuItemsProps } from './components';
+export { SpectrogramSettingsModal } from './components';
+export type { SpectrogramSettingsModalProps } from './components';
+export type { TrackMenuItem } from './components';
