+++ packages/spectrogram/src/styled.d.ts
@@ -1,0 +1,9 @@
+import 'styled-components';
+
+declare module 'styled-components' {
+  export interface DefaultTheme {
+    timescaleBackgroundColor?: string;
+    textColor?: string;
+    waveProgressColor?: string;
+  }
+}
