+++ packages/ui-components/src/components/index.tsx
@@ -22,5 +22,4 @@ export * from './Track';
 export * from './TrackControls';
 export * from './TrackControlsWithDelete';
 export * from './TrackMenu';
-export * from './SpectrogramSettingsModal';
 export { StyledTimeScale } from './TimeScale';
